+++ hasher-matcher-actioner/mypy.ini
@@ -1,2 +1,23 @@
 [mypy-threatexchange.hashing.*]
 ignore_missing_imports = True
+
+[mypy-threatexchange.signal_type.pdq]
+ignore_missing_imports = True
+
+[mypy-threatexchange.cli.dataset.simple_serialization]
+ignore_missing_imports = True
+
+[mypy-threatexchange.api]
+ignore_missing_imports = True
+
+[mypy-threatexchange.threat_updates]
+ignore_missing_imports = True
+
+[mypy-threatexchange.signal_type.pdq_index]
+ignore_missing_imports = True
+
+[mypy-bottle.*]
+ignore_missing_imports = True
+
+[mypy-apig_wsgi.*]
+ignore_missing_imports = True
