+++ hasher-matcher-actioner/requirements-dev.txt
@@ -1,2 +1,3 @@
+-r requirements.txt
 pytest==6.2.1
 mypy==0.812
