+++ hasher-matcher-actioner/tests/test_sample.py
@@ -1,2 +1,2 @@
 def test_sample():
-    assert True == False
+    assert True == True
